requirements.txt
numpy==1.26.4
streamlit==1.45.0
PyPDF2==3.0.1
pinecone==6.0.2
langchain==0.3.25
transformers==4.51.3
torch==2.5.0
python-dotenv==1.1.0
groq==0.24.0